+++ passes/R008/testdata/src/a_fixed/main.go
@@ -16,13 +16,13 @@ func f() {
 
 	// Failing
 
-	d.SetPartial("test") // want "deprecated d.SetPartial"
+	// want "deprecated d.SetPartial"
 
 	fResourceFunc(&d, nil)
 }
 
 func fResourceFunc(d *schema.ResourceData, meta interface{}) error {
-	d.SetPartial("test") // want "deprecated d.SetPartial"
+	// want "deprecated d.SetPartial"
 
 	return nil
 }